test: split live snapshot/event tests [OPE-506]
